Update 2019.md
A bit more frank about the status of the Neo4j project as we left it at the end of GSoC 2017. I'm not sure about the Bluegenes part, but I left it in there. (I don't think we want to spend a lot of time developing client-side code for a backing store that never gets into production.)